// src/config/features/types.ts
export type HeaderLinksType = 'sports' | 'in-play' | 'virtuals-leap' | 'virtuals-nyx' | 'casino' | 'live-casino' | 'mcbookie-blog' | 'star-news' | 'lottery' | 'rhino-promotions';
export type HeaderLinksMenuType = 'sports' | 'in-play' | 'virtuals' | 'casino' | 'live-casino' | 'mcbookie-blog' | 'star-news' | 'lottery' | 'rhino-promotions';